Various fixes for accurate implementation of PolyBench C

+++ benchmarks/autoscheduler_benchmarks/polybench/LARGE/function_floyd_warshall_LARGE/function_floyd_warshall_LARGE_wrapper.cpp
@@ -12,9 +12,9 @@ using namespace std;
 int main(int, char **argv)
 {
 
-	double *b_D = (double*)malloc(2800*2800*sizeof(double));
-	parallel_init_buffer(b_D, 2800*2800, (double)19);
-	Halide::Buffer<double> buf00(b_D, 2800, 2800);
+	int *b_D = (int*)malloc(2800*2800*sizeof(int));
+	parallel_init_buffer(b_D, 2800*2800, (int)19);
+	Halide::Buffer<int> buf00(b_D, 2800, 2800);
