+++ benchmarks/autoscheduler_benchmarks/polybench/MINI/function_floyd_warshall_MINI/function_floyd_warshall_MINI_wrapper.cpp
@@ -12,9 +12,9 @@ using namespace std;
 int main(int, char **argv)
 {
 
-	double *b_D = (double*)malloc(60*60*sizeof(double));
-	parallel_init_buffer(b_D, 60*60, (double)19);
-	Halide::Buffer<double> buf00(b_D, 60, 60);
+	int *b_D = (int*)malloc(60*60*sizeof(int));
+	parallel_init_buffer(b_D, 60*60, (int)19);
+	Halide::Buffer<int> buf00(b_D, 60, 60);
